* Updated the FreeBSD compatibility layer for network drivers to FreeBSD 7   (RELENG_7_BP). * There are many white spots, though, most notable PCI MSI(-X) support, and   jumbo frames. * Fixed removing interrupts for the INTR_FAST case. Since FreeBSD 7 added   a new interrupt "filter" mechanism, we can finally report if the interface   was handled by a device or not (though only very few devices support this   yet). * Updated the 3com, rtl8139, e1000, and via_rhine drivers to the latest code   base. They all compile, but I haven't tested them after the changes yet!
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@22991 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/libs/compat/freebsd_network/if.c
@@ -319,7 +319,7 @@ ether_ifdetach(struct ifnet *ifp)
 
 
 int
-ether_ioctl(struct ifnet *ifp, int command, caddr_t data)
+ether_ioctl(struct ifnet *ifp, u_long command, caddr_t data)
 {
 	struct ifreq *ifr = (struct ifreq *)data;
